backend/requirements.txt
Flask==2.2.2
werkzeug==2.2.2
gunicorn==20.1.0

pytest==7.1.3
pytest-cov==4.0.0
flake8==5.0.4
psycopg2==2.9.3

SQLAlchemy==1.4.42
Flask-SQLAlchemy==3.0.2
Flask-Cors==3.0.10

python-dotenv
